+++ lib/src/platform/reader_platform.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
@@ -17,10 +18,18 @@ abstract class ReaderPlatform extends PlatformInterface {
   Future<void> setKeepScreenOn(bool enabled) {
     throw UnimplementedError('setKeepScreenOn() has not been implemented.');
   }
+
+  bool get supportsKeepScreenOn => false;
 }
 
 class MethodChannelReaderPlatform extends ReaderPlatform {
   static const MethodChannel _channel = MethodChannel('novel_reader_ui/system');
+
+  @override
+  bool get supportsKeepScreenOn =>
+      !kIsWeb &&
+      (defaultTargetPlatform == TargetPlatform.android ||
+          defaultTargetPlatform == TargetPlatform.windows);
 
   @override
   Future<void> setKeepScreenOn(bool enabled) {